In src/SpikingNeuralNets.jl, src/SimpleSNN.jl, src/LIF.jl: Modified spike history S to be a vector of dictionaries S[neuron][time] to enable multiple spikes within a single timestep.
In src/LIF.jl: Removed the multiplication of each channel's varFns with dt to absorb dt into the parameters of each varFn, reducing the number of computations per timestep.

In test/RecurrentCorticalNetwork.jl: implemented singleneurontest and synapsetest functions to verify the single neuron/synapse properties of the network. Implemented an iterator for the input stimulus to save memory.

Added test/reduced-test.jl, the beginnings of a reduced two-variable model of a SNN.